extracted model spec from main spec
