+++ lib/tddetective-model.js
@@ -71,7 +71,7 @@ export default class TddetectiveModel {
   }
 
   _hasSpecFileName(aClassName) {
-    var path = this._getSpecPath() + '/' + aClassName.toLowerCase() + '-spec.js'
+    var path = this._getSpecPath() + '/' + aClassName.toLowerCase() + '_spec.rb'
     if (fs.existsSync(path)) {
       specFileNames.push(path);
     }
